bot/pyproject.toml
[tool.poetry]
name = "w24ok-bot-bot"
version = "0.1.0"
description = ""
authors = ["ArthurGolubev <gg.arthur.go@gmail.com>"]
readme = "README.md"
packages = [{include = "w24ok_bot_bot"}]

[tool.poetry.dependencies]
python = "^3.10"
loguru = "^0.6.0"
python-telegram-bot = "^20.1"
notifiers = "^1.3.3"
psycopg2-binary = "^2.9.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
